refactored map controller into package/map
